# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMY 施工管理 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      parameters:
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/ProjectStatus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /projects/{id}/ledger:
    get:
      operationId: getProjectLedger
      tags: [projects]
      summary: Construction cost ledger summary for a project (planned vs actual)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectLedger" }
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      parameters:
        - in: query
          name: projectId
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Quote" }
    post:
      operationId: createQuote
      tags: [quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateQuoteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quote" }
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quote" }
    patch:
      operationId: updateQuote
      tags: [quotes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateQuoteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quote" }
    delete:
      operationId: deleteQuote
      tags: [quotes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
  /schedule:
    get:
      operationId: listScheduleEntries
      tags: [schedule]
      parameters:
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScheduleEntry" }
    post:
      operationId: createScheduleEntry
      tags: [schedule]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateScheduleEntryBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleEntry" }
  /schedule/{id}:
    patch:
      operationId: updateScheduleEntry
      tags: [schedule]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateScheduleEntryBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleEntry" }
    delete:
      operationId: deleteScheduleEntry
      tags: [schedule]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ProjectStatus:
      type: string
      enum: [estimating, contracted, in_progress, completed, archived]
    CostCategory:
      type: string
      enum: [material, subcontract, labor, expense, other]
...
    Project:
      type: object
      required: [id, name, status, customerId, customerName, contractAmount, plannedCost, actualCost, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string, nullable: true, description: "案件番号" }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        customerId: { type: string }
        customerName: { type: string }
        siteAddress: { type: string, nullable: true }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }
        contractAmount: { type: number, description: "契約金額 (税抜, 円)" }
        plannedCost: { type: number, description: "予算原価合計" }
        actualCost: { type: number, description: "実績原価合計" }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateProjectBody:
      type: object
      required: [name, customerId, status]
      properties:
        name: { type: string }
        code: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        customerId: { type: string }
        siteAddress: { type: string, nullable: true }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }
        contractAmount: { type: number }
        notes: { type: string, nullable: true }
    UpdateProjectBody:
      type: object
      properties:
        name: { type: string }
        code: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        customerId: { type: string }
        siteAddress: { type: string, nullable: true }
        startDate: { type: string, format: date, nullable: true }
        endDate: { type: string, format: date, nullable: true }
        contractAmount: { type: number }
        notes: { type: string, nullable: true }
    LineItem:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        description: { type: string }
        unit: { type: string, nullable: true }
        quantity: { type: number }
        unitPrice: { type: number }
    Quote:
      type: object
      required: [id, projectId, quoteNumber, issueDate, items, subtotal, tax, total, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        projectName: { type: string }
        quoteNumber: { type: string }
        issueDate: { type: string, format: date }
        validUntil: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/LineItem" }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        createdAt: { type: string, format: date-time }
    CreateQuoteBody:
      type: object
      required: [projectId, quoteNumber, issueDate, items]
      properties:
        projectId: { type: string }
        quoteNumber: { type: string }
        issueDate: { type: string, format: date }
        validUntil: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/LineItem" }
...
    CostEntry:
      type: object
      required: [id, projectId, category, description, plannedAmount, actualAmount, entryDate, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        projectName: { type: string }
        category: { $ref: "#/components/schemas/CostCategory" }
        description: { type: string, description: "工事項目 / 摘要" }
        vendor: { type: string, nullable: true, description: "仕入先 / 外注先" }
        plannedAmount: { type: number, description: "予算原価 (円)" }
        actualAmount: { type: number, description: "実績原価 (円)" }
        entryDate: { type: string, format: date }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    ScheduleEntry:
      type: object
      required: [id, projectId, projectName, staffId, staffName, date, task, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        projectName: { type: string }
        staffId: { type: string }
        staffName: { type: string }
        date: { type: string, format: date }
        task: { type: string }
        startTime: { type: string, nullable: true, description: "HH:MM" }
        endTime: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateScheduleEntryBody:
      type: object
      required: [projectId, staffId, date, task]
      properties:
        projectId: { type: string }
        staffId: { type: string }
        date: { type: string, format: date }
        task: { type: string }
        startTime: { type: string, nullable: true }
        endTime: { type: string, nullable: true }
        notes: { type: string, nullable: true }
...
    ProjectLedger:
      type: object
      required: [projectId, contractAmount, plannedCost, actualCost, grossProfit, byCategory, entries]
      properties:
        projectId: { type: string }
        projectName: { type: string }
        contractAmount: { type: number }
        plannedCost: { type: number }
        actualCost: { type: number }
        grossProfit: { type: number, description: "粗利 = 契約金額 - 実績原価" }
        grossProfitRate: { type: number, description: "粗利率 (%)" }
        byCategory:
          type: array
          items: { $ref: "#/components/schemas/LedgerCategoryRow" }
        entries:
          type: array
          items: { $ref: "#/components/schemas/CostEntry" }
    LedgerCategoryRow:
      type: object
      required: [category, plannedAmount, actualAmount]
      properties:
        category: { $ref: "#/components/schemas/CostCategory" }
        plannedAmount: { type: number }
        actualAmount: { type: number }
...