# lib/api-zod/src/generated/types/createQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMY 施工管理 API
 * OpenAPI spec version: 0.1.0
 */
import type { LineItem } from "./lineItem";

export interface CreateQuoteBody {
  projectId: string;
  quoteNumber: string;
  issueDate: Date;
  validUntil?: Date | null;
  notes?: string | null;
  items: LineItem[];
}