# lib/api-zod/src/generated/types/createScheduleEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMY 施工管理 API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateScheduleEntryBody {
  projectId: string;
  staffId: string;
  date: Date;
  task: string;
  startTime?: string | null;
  endTime?: string | null;
  notes?: string | null;
}