# lib/api-zod/src/generated/types/ledgerCategoryRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMY 施工管理 API
 * OpenAPI spec version: 0.1.0
 */
import type { CostCategory } from "./costCategory";

export interface LedgerCategoryRow {
  category: CostCategory;
  plannedAmount: number;
  actualAmount: number;
}