# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMY 施工管理 API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from "./projectStatus";

export interface UpdateProjectBody {
  name?: string;
  code?: string | null;
  status?: ProjectStatus;
  customerId?: string;
  siteAddress?: string | null;
  startDate?: Date | null;
  endDate?: Date | null;
  contractAmount?: number;
  notes?: string | null;
}